+++ scim2_models/scim_object.py
@@ -17,6 +17,7 @@ from .annotations import Required
 from .base import BaseModel
 from .context import Context
 from .path import URN
+from .path import Path
 
 if TYPE_CHECKING:
     pass
@@ -105,6 +106,8 @@ class ScimObject(BaseModel, metaclass=ScimMetaclass):
     def _prepare_model_dump(
         self,
         scim_ctx: Context | None = Context.DEFAULT,
+        attributes: list[str | Path[Any]] | None = None,
+        excluded_attributes: list[str | Path[Any]] | None = None,
         **kwargs: Any,
     ) -> dict[str, Any]:
         kwargs.setdefault("context", {}).setdefault("scim", scim_ctx)
@@ -113,12 +116,21 @@ class ScimObject(BaseModel, metaclass=ScimMetaclass):
             kwargs.setdefault("exclude_none", True)
             kwargs.setdefault("by_alias", True)
 
+        if attributes:
+            kwargs["context"]["scim_attributes"] = [str(a) for a in attributes]
+        if excluded_attributes:
+            kwargs["context"]["scim_excluded_attributes"] = [
+                str(a) for a in excluded_attributes
+            ]
+
         return kwargs
 
     def model_dump(
         self,
         *args: Any,
         scim_ctx: Context | None = Context.DEFAULT,
+        attributes: list[str | Path[Any]] | None = None,
+        excluded_attributes: list[str | Path[Any]] | None = None,
         **kwargs: Any,
     ) -> dict[str, Any]:
         """Create a model representation that can be included in SCIM messages by using Pydantic :code:`BaseModel.model_dump`.
@@ -126,8 +138,18 @@ class ScimObject(BaseModel, metaclass=ScimMetaclass):
         :param scim_ctx: If a SCIM context is passed, some default values of
             Pydantic :code:`BaseModel.model_dump` are tuned to generate valid SCIM
             messages. Pass :data:`None` to get the default Pydantic behavior.
+        :param attributes: A multi-valued list of strings indicating the names of resource
+            attributes to return in the response, overriding the set of attributes that
+            would be returned by default. Invalid values are ignored.
+        :param excluded_attributes: A multi-valued list of strings indicating the names of resource
+            attributes to be removed from the default set of attributes to return. Invalid values are ignored.
         """
-        dump_kwargs = self._prepare_model_dump(scim_ctx, **kwargs)
+        dump_kwargs = self._prepare_model_dump(
+            scim_ctx,
+            attributes=attributes,
+            excluded_attributes=excluded_attributes,
+            **kwargs,
+        )
         if scim_ctx:
             dump_kwargs.setdefault("mode", "json")
         return super(BaseModel, self).model_dump(*args, **dump_kwargs)
@@ -136,6 +158,8 @@ class ScimObject(BaseModel, metaclass=ScimMetaclass):
         self,
         *args: Any,
         scim_ctx: Context | None = Context.DEFAULT,
+        attributes: list[str | Path[Any]] | None = None,
+        excluded_attributes: list[str | Path[Any]] | None = None,
         **kwargs: Any,
     ) -> str:
         """Create a JSON model representation that can be included in SCIM messages by using Pydantic :code:`BaseModel.model_dump_json`.
@@ -143,6 +167,16 @@ class ScimObject(BaseModel, metaclass=ScimMetaclass):
         :param scim_ctx: If a SCIM context is passed, some default values of
             Pydantic :code:`BaseModel.model_dump` are tuned to generate valid SCIM
             messages. Pass :data:`None` to get the default Pydantic behavior.
+        :param attributes: A multi-valued list of strings indicating the names of resource
+            attributes to return in the response, overriding the set of attributes that
+            would be returned by default. Invalid values are ignored.
+        :param excluded_attributes: A multi-valued list of strings indicating the names of resource
+            attributes to be removed from the default set of attributes to return. Invalid values are ignored.
         """
-        dump_kwargs = self._prepare_model_dump(scim_ctx, **kwargs)
+        dump_kwargs = self._prepare_model_dump(
+            scim_ctx,
+            attributes=attributes,
+            excluded_attributes=excluded_attributes,
+            **kwargs,
+        )
         return super(BaseModel, self).model_dump_json(*args, **dump_kwargs)